pyproject.toml
[project]
name = "codebase"
version = "0.1.0"
description = "MadyPro - A Modern App Develpment Project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic==1.16.1",
    "aniso8601==7.0.0",
    "bcrypt==4.3.0",
    "blinker==1.9.0",
    "boto3==1.42.62",
    "botocore==1.42.62",
    "certifi==2024.8.30",
    "charset-normalizer==3.4.0",
    "click==8.2.1",
    "coverage==7.8.2",
    "dnspython==2.7.0",
    "email-validator==2.2.0",
    "flask==3.1.1",
    "flask-cors==6.0.0",
    "flask-graphql==2.0.1",
    "flask-login==0.6.3",
    "flask-migrate==4.1.0",
    "flask-restful==0.3.10",
    "flask-sqlalchemy==3.1.1",
    "flask-wtf==1.2.2",
    "graphene==2.1.9",
    "graphene-sqlalchemy==2.1.0",
    "graphql-core==2.3.2",
    "graphql-relay==2.0.1",
    "graphql-server-core==1.2.0",
    "greenlet==3.2.2",
    "gunicorn==25.1.0",
    "idna==3.10",
    "iniconfig==2.1.0",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "jmespath==1.1.0",
    "mako==1.3.10",
    "markupsafe==3.0.2",
    "matplotlib>=3.10.8",
    "packaging==25.0",
    "pillow==11.0.0",
    "pluggy==1.6.0",
    "promise==2.3",
    "pymysql==1.1.2",
    "pytest==8.3.5",
    "pytest-cov==6.1.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "pytz==2025.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "s3transfer==0.16.0",
    "setuptools==75.5.0",
    "six==1.17.0",
    "sqlalchemy==2.0.41",
    "tqdm==4.67.1",
    "typing-extensions==4.13.2",
    "urllib3==2.2.3",
    "werkzeug==3.1.3",
    "wtforms==3.2.1",
]
